+++ src/modules/search/services/index.ts
@@ -1,1 +1,2 @@
 export { KeywordSearchService } from './KeywordSearchService';
+export { SemanticSearchService } from './SemanticSearchService';